backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
gTTS
dnspython
certifi
requests
pydantic
sendgrid
twilio
